Phase 2 PR 9: the astro client — retries, a breaker, and honest doubles
Task 2.15. Tier 1+, not the cached Tier 0: the contract changed and the
client regenerated.

## The retry transport excluded every call this client makes

The shared transport from Phase 0 refuses to retry POST, with a good
comment explaining why: a replayed POST is a duplicated write, and nobody
notices until reconciliation.

Every astro-service route is a POST — birth data is far too large for a
query string — so the spec's "two retries on 5xx" was, in practice,
retrying nothing.

Resolved by making idempotency an explicit per-call declaration rather
than weakening the method rule for everyone. `MarkIdempotent(ctx)` is
stated at each astro call site, and it is true by TOPOLOGY rather than by
hope: astro-service has no database at all — invariant 2 — so there is no
write for a replay to duplicate. A reviewer adding a stateful endpoint
has to look at the claim to make it.

Proven by removing the opt-in: the retry test fails, the POST is
attempted once.

## The breaker, and an ordering claim that no test pinned

Retrying is right for a blip and wrong for an outage. With astro down,
every request spends its full timeout and both retries — ten seconds
times three — holding a goroutine and a connection while a browser waits.
The breaker turns that into an immediate failure so the chart service can
serve the cached chart, which is the property the spec asks for.

It sits OUTSIDE the retry transport, so one logical call is one
observation. Inside, each retry counts separately and the circuit trips
about three times sooner than configured — five consecutive failures
means five failed operations, not five failed packets.

I wrote that reasoning as a comment, then swapped the ordering to check:
the entire suite still passed. The claim was unpinned. There is now a
test that counts CALLS to open rather than failures, and with the breaker
moved inside it reports "the circuit opened after 2 calls" against a
threshold of 5.

That is the third comment this phase asserting something no test
enforced. Each time the fix has been a test, not a better comment.

## Failure classification, because the cache fallback depends on it

  5xx or a dead connection  ErrAstroUnavailable  → serve the cached chart
  4xx                       ErrAstroRejected     → do NOT serve a cache

A 4xx is our own bad request. Serving a cached chart there would hide a
bug that produces wrong charts for every new profile from then on. A 4xx
also does not trip the breaker: one caller's bad request taking astro
down for everybody turns a bug into an outage.

## float32 on the wire, measured before deciding

The generated client typed latitude and longitude as float32, because
Pydantic emits a bare `{"type": "number"}` and oapi-codegen maps that to
single precision.

Measured rather than assumed: the worst case across the coordinate range
is 24 milliarcseconds, about 0.76 m on the ground, and six orders of
magnitude below a pada boundary. It does not matter for degrees.

Fixed anyway, at the source, because the default applies to EVERY float
in the contract. A Julian day is 2451545.0 before the decimal point;
float32 carries about seven significant digits in total, so such a field
would silently lose the time of day. Declaring `format: double` once
means the next numeric field cannot be narrowed by accident.

13 client tests. task verify green. Both Go suites green with -race.

Beads: n/a — Tier 0 classifier, treated as Tier 1 (contract + client)
Change Contract: docs/specs/PHASE-02-ASTROLOGY-ENGINE.md task 2.15, §8

+++ services/api/internal/platform/clients/astroclient/client.gen.go
@@ -199,10 +199,10 @@ func (e YogaResultStrength) Valid() bool {
 
 // AscendantPosition defines model for AscendantPosition.
 type AscendantPosition struct {
-	Degree float32 `json:"degree"`
+	Degree float64 `json:"degree"`
 
 	// Longitude Sidereal ecliptic longitude
-	Longitude float32 `json:"longitude"`
+	Longitude float64 `json:"longitude"`
 	Nakshatra string  `json:"nakshatra"`
 	Pada      int     `json:"pada"`
 	Sign      string  `json:"sign"`
@@ -222,8 +222,8 @@ type AscendantPosition struct {
 type BirthData struct {
 	Ayanamsa     *BirthDataAyanamsa     `json:"ayanamsa,omitempty"`
 	HouseSystem  *BirthDataHouseSystem  `json:"house_system,omitempty"`
-	Latitude     float32                `json:"latitude"`
-	Longitude    float32                `json:"longitude"`
+	Latitude     float64                `json:"latitude"`
+	Longitude    float64                `json:"longitude"`
 	TimeAccuracy *BirthDataTimeAccuracy `json:"time_accuracy,omitempty"`
 
 	// UtcInstant Birth instant in UTC. Must be timezone-aware.
@@ -244,7 +244,7 @@ type BirthDataTimeAccuracy string
 // ChartMeta Provenance. Every one of these fields answers "why did it say that?".
 type ChartMeta struct {
 	Ayanamsa          string                      `json:"ayanamsa"`
-	AyanamsaValue     float32                     `json:"ayanamsa_value"`
+	AyanamsaValue     float64                     `json:"ayanamsa_value"`
 	CalculationSystem *ChartMetaCalculationSystem `json:"calculation_system,omitempty"`
 	ComputedAt        time.Time                   `json:"computed_at"`
 	EngineVersion     string                      `json:"engine_version"`
@@ -323,12 +323,12 @@ type DashaRequest struct {
 	MaxLevel     *int      `json:"max_level,omitempty"`
 
 	// MoonLongitude Sidereal ecliptic longitude
-	MoonLongitude float32 `json:"moon_longitude"`
+	MoonLongitude float64 `json:"moon_longitude"`
 }
 
 // DashaResponse defines model for DashaResponse.
 type DashaResponse struct {
-	BalanceAtBirthDays float32          `json:"balance_at_birth_days"`
+	BalanceAtBirthDays float64          `json:"balance_at_birth_days"`
 	Periods            []DashaPeriodOut `json:"periods"`
 }
 
@@ -365,14 +365,14 @@ type HousePosition struct {
 // PlanetPosition defines model for PlanetPosition.
 type PlanetPosition struct {
 	Aspects      *[]int                `json:"aspects,omitempty"`
-	Degree       float32               `json:"degree"`
+	Degree       float64               `json:"degree"`
 	Dignity      PlanetPositionDignity `json:"dignity"`
 	House        int                   `json:"house"`
 	IsCombust    bool                  `json:"is_combust"`
 	IsRetrograde bool                  `json:"is_retrograde"`
 
 	// Longitude Sidereal ecliptic longitude
-	Longitude      float32 `json:"longitude"`
+	Longitude      float64 `json:"longitude"`
 	Nakshatra      string  `json:"nakshatra"`
 	NakshatraIndex int     `json:"nakshatra_index"`
 	Pada           int     `json:"pada"`
@@ -381,7 +381,7 @@ type PlanetPosition struct {
 
 	// SignIndex 0 = Aries
 	SignIndex int     `json:"sign_index"`
-	Speed     float32 `json:"speed"`
+	Speed     float64 `json:"speed"`
 }
 
 // PlanetPositionDignity defines model for PlanetPosition.Dignity.
@@ -401,13 +401,13 @@ type SadeSatiResultCurrentPhase string
 
 // TransitPosition defines model for TransitPosition.
 type TransitPosition struct {
-	Degree             float32 `json:"degree"`
+	Degree             float64 `json:"degree"`
 	HouseFromAscendant *int    `json:"house_from_ascendant"`
 	HouseFromMoon      int     `json:"house_from_moon"`
 	IsRetrograde       bool    `json:"is_retrograde"`
 
 	// Longitude Sidereal ecliptic longitude
-	Longitude float32 `json:"longitude"`
+	Longitude float64 `json:"longitude"`
 	Planet    string  `json:"planet"`
 	Sign      string  `json:"sign"`
 
@@ -432,7 +432,7 @@ type TransitRequestAyanamsa string
 type TransitResponse struct {
 	At            time.Time         `json:"at"`
 	Ayanamsa      string            `json:"ayanamsa"`
-	AyanamsaValue float32           `json:"ayanamsa_value"`
+	AyanamsaValue float64           `json:"ayanamsa_value"`
 	SadeSati      SadeSatiResult    `json:"sade_sati"`
 	Transits      []TransitPosition `json:"transits"`
 }
